[Feat] SwiftDataSource 구현
